refactor: 프론트엔드 폴더 구조 변경
