+++ frontend/src/pages/Mainpage/handlers/MainpageHandlers.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
-import { useNotification } from '../Notification/Notification';
+import { useNotification } from '../../Notification/Notification';
 import {useMemberHandlers} from './MemberHandlers.jsx';
 import { usePenaltyHandlers } from './PenaltyHandlers.jsx';
