+++ frontend/src/pages/Mainpage/handlers/PenaltyHandlers.jsx
